requirements.txt
Django==1.11
flake8==3.6.0
mccabe==0.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
djangorestframework==3.6.3
